test: Update OAuthRequestValidatorTest to test new implementation rather than legacy
