+++ channel-azure-boards/src/test/java/com/synopsys/integration/alert/channel/azure/boards/oauth/OAuthRequestValidatorTest.java
@@ -1,0 +1,46 @@
+package com.synopsys.integration.alert.channel.azure.boards.oauth;
+
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.util.UUID;
+
+import org.junit.jupiter.api.Test;
+
+class OAuthRequestValidatorTest {
+
+    @Test
+    void testAddKey() {
+        OAuthRequestValidator oAuthRequestValidator = new OAuthRequestValidator();
+        UUID requestKey = oAuthRequestValidator.generateRequestKey();
+        oAuthRequestValidator.addAuthorizationRequest(requestKey, UUID.randomUUID());
+        assertTrue(oAuthRequestValidator.hasRequestKey(requestKey));
+    }
+
+    @Test
+    void addNullKeyTest() {
+        OAuthRequestValidator oAuthRequestValidator = new OAuthRequestValidator();
+        oAuthRequestValidator.addAuthorizationRequest(null, UUID.randomUUID());
+        assertFalse(oAuthRequestValidator.hasRequestKey(null));
+    }
+
+    @Test
+    void testRemoveKey() {
+        OAuthRequestValidator oAuthRequestValidator = new OAuthRequestValidator();
+        UUID requestKey = oAuthRequestValidator.generateRequestKey();
+        oAuthRequestValidator.addAuthorizationRequest(requestKey, UUID.randomUUID());
+        oAuthRequestValidator.removeAuthorizationRequest(requestKey);
+        assertFalse(oAuthRequestValidator.hasRequestKey(requestKey));
+    }
+
+    @Test
+    void removeNullKeyTest() {
+        OAuthRequestValidator oAuthRequestValidator = new OAuthRequestValidator();
+        UUID requestKey = oAuthRequestValidator.generateRequestKey();
+        oAuthRequestValidator.addAuthorizationRequest(requestKey, UUID.randomUUID());
+        oAuthRequestValidator.removeAuthorizationRequest(null);
+        assertFalse(oAuthRequestValidator.hasRequestKey(null));
+        assertTrue(oAuthRequestValidator.hasRequestKey(requestKey));
+        assertTrue(oAuthRequestValidator.hasRequests());
+    }
+}
